tests: move RedisServer to a common module
