+++ tests/test_basic.rs
@@ -4,155 +4,14 @@ extern crate net2;
 
 use redis::{Commands, PipelineCommands};
 
-use std::fs;
-use std::env;
-use std::process;
 use std::thread::{spawn, sleep};
 use std::time::Duration;
 use std::collections::{HashMap, HashSet};
 use std::collections::{BTreeSet,BTreeMap};
 
-use std::path::PathBuf;
+mod common;
 
-#[derive(PartialEq)]
-enum ServerType {
-    Tcp,
-    Unix,
-}
-
-pub struct RedisServer {
-    pub process: process::Child,
-    addr: redis::ConnectionAddr,
-}
-
-impl ServerType {
-    fn get_intended() -> ServerType {
-        match env::var("REDISRS_SERVER_TYPE").ok().as_ref().map(|x| &x[..]) {
-            Some("tcp") => ServerType::Tcp,
-            Some("unix") => ServerType::Unix,
-            val => {
-                panic!("Unknown server type {:?}", val);
-            }
-        }
-    }
-}
-
-impl RedisServer {
-    pub fn new() -> RedisServer {
-        let server_type = ServerType::get_intended();
-        let mut cmd = process::Command::new("redis-server");
-        cmd.stdout(process::Stdio::null())
-            .stderr(process::Stdio::null());
-
-        let addr = match server_type {
-            ServerType::Tcp => {
-                // this is technically a race but we can't do better with
-                // the tools that redis gives us :(
-                let listener = net2::TcpBuilder::new_v4()
-                    .unwrap()
-                    .reuse_address(true)
-                    .unwrap()
-                    .bind("127.0.0.1:0")
-                    .unwrap()
-                    .listen(1)
-                    .unwrap();
-                let server_port = listener.local_addr().unwrap().port();
-                cmd.arg("--port")
-                    .arg(server_port.to_string())
-                    .arg("--bind")
-                    .arg("127.0.0.1");
-                redis::ConnectionAddr::Tcp("127.0.0.1".to_string(), server_port)
-            }
-            ServerType::Unix => {
-                let (a, b) = rand::random::<(u64, u64)>();
-                let path = format!("/tmp/redis-rs-test-{}-{}.sock", a, b);
-                cmd.arg("--port")
-                    .arg("0")
-                    .arg("--unixsocket")
-                    .arg(&path);
-                redis::ConnectionAddr::Unix(PathBuf::from(&path))
-            }
-        };
-
-        let process = cmd.spawn().unwrap();
-        RedisServer {
-            process: process,
-            addr: addr,
-        }
-    }
-
-    pub fn wait(&mut self) {
-        self.process.wait().unwrap();
-    }
-
-    pub fn get_client_addr(&self) -> &redis::ConnectionAddr {
-        &self.addr
-    }
-}
-
-impl Drop for RedisServer {
-    fn drop(&mut self) {
-        let _ = self.process.kill();
-        let _ = self.process.wait();
-        match *self.get_client_addr() {
-            redis::ConnectionAddr::Unix(ref path) => {
-                fs::remove_file(&path).ok();
-            }
-            _ => {}
-        }
-    }
-}
-
-pub struct TestContext {
-    pub server: RedisServer,
-    pub client: redis::Client,
-}
-
-impl TestContext {
-    fn new() -> TestContext {
-        let server = RedisServer::new();
-
-        let client = redis::Client::open(redis::ConnectionInfo {
-                addr: Box::new(server.get_client_addr().clone()),
-                db: 0,
-                passwd: None,
-            })
-            .unwrap();
-        let con;
-
-        let millisecond = Duration::from_millis(1);
-        loop {
-            match client.get_connection() {
-                Err(err) => {
-                    if err.is_connection_refusal() {
-                        sleep(millisecond);
-                    } else {
-                        panic!("Could not connect: {}", err);
-                    }
-                }
-                Ok(x) => {
-                    con = x;
-                    break;
-                }
-            }
-        }
-        redis::cmd("FLUSHDB").execute(&con);
-
-        TestContext {
-            server: server,
-            client: client,
-        }
-    }
-
-    fn connection(&self) -> redis::Connection {
-        self.client.get_connection().unwrap()
-    }
-
-    fn pubsub(&self) -> redis::PubSub {
-        self.client.get_pubsub().unwrap()
-    }
-}
-
+use common::*;
 
 #[test]
 fn test_parse_redis_url() {
